+++ R/iemisc.R
@@ -4,13 +4,14 @@
 #' coefficient of variation (CV), approximate and relative error, range,
 #' harmonic mean, geometric mean], engineering economics (benefit-cost,
 #' future value, present value, annual value, gradients, interest, periods,
-#' etc.), geometry (sphere volume and right triangle), environmental/water
-#' resources engineering (Manning's n, Gauckler-Manning-Strickler equation for
-#' geometric cross-sections), a version of linear interpolation for use with
-#' NAs, & GNU Octave/MATLAB compatible size, numel, and length_octave functions.
+#' etc.), geometry (sphere volume and right triangle), civil & environmental/
+#' water resources engineering (Normal Strength Concrete Mix Design, Manning's
+#' n, Gauckler-Manning-Strickler equation for geometric cross-sections), a
+#' version of linear interpolation for use with NAs, & GNU Octave/MATLAB
+#' compatible size, numel, and length_octave functions.
 #'
 #' @docType package
 #' @name iemisc
-#' @import data.table stats
-#' @importFrom pracma isempty nthroot interp1
+#' @import data.table stats iemiscdata import zoo gsubfn fpCompare
+#' @importFrom pracma isempty nthroot interp1 ones
 NULL